Add duration meter and notification for SlipTime
